+++ index.js
@@ -5,6 +5,7 @@ export { withStripes } from './src/StripesContext';
 export { default as Pluggable } from './src/Pluggable';
 export { setServicePoints, setCurServicePoint } from './src/loginServices';
 export { default as IfInterface } from './src/components/IfInterface';
+export { default as IfPermission } from './src/components/IfPermission';
 export { default as TitleManager } from './src/components/TitleManager';
 export { default as HandlerManager } from './src/components/HandlerManager';
 export { default as coreEvents } from './src/events';