+++ src/components/ForgotPassword/ForgotPasswordWrapper.js
@@ -5,11 +5,6 @@ import ForgotPasswordCtrl from './ForgotPasswordCtrl';
 
 class ForgotPasswordWrapper extends Component {
   static propTypes = {
-    mutator: PropTypes.shape({
-      searchUsername: PropTypes.shape({
-        POST: PropTypes.func.isRequired,
-      }).isRequired,
-    }).isRequired,
     stripes: PropTypes.object.isRequired,
   };